# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VolunteerHub API",
    "version": "1.0.0",
    "description": "API documentation for VolunteerHub backend"
  },
  "servers": [
    { "url": "http://localhost:5000/api", "description": "Local API" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "bio": { "type": "string", "nullable": true },
          "avatarUrl": { "type": "string", "nullable": true },
          "role": { "type": "string", "enum": ["VOLUNTEER", "MANAGER", "ADMIN"] },
          "isLocked": { "type": "boolean" },
          "pushNotifications": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "AuthLoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "AuthRegisterRequest": {
        "type": "object",
        "required": ["email", "password", "firstName", "lastName"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "role": {
            "type": "string",
            "enum": ["VOLUNTEER", "MANAGER"],
            "description": "User role during registration. Defaults to VOLUNTEER if not provided."
          }
        }
      },
      "LockRequest": {
        "type": "object",
        "required": ["isLocked"],
        "properties": { "isLocked": { "type": "boolean" } }
      },
      "UpdateMeRequest": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "bio": { "type": "string" },
          "avatarUrl": { "type": "string" },
          "pushNotifications": { "type": "boolean" },
          "currentPassword": { "type": "string" },
          "newPassword": { "type": "string" }
        }
      },
      "UploadAvatarRequest": {
        "type": "object",
        "properties": {
          "dataUrl": { "type": "string", "description": "data:image/png;base64,..." },
          "base64": { "type": "string" },
          "contentType": { "type": "string", "example": "image/png" }
        }
      },
      "Event": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "location": { "type": "string" },
          "category": { "type": "string" },
          "maxParticipants": { "type": "integer" },
          "currentParticipants": { "type": "integer" },
          "thumbnailUrl": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["DRAFT","SUBMITTED","PENDING_APPROVAL","APPROVED","REJECTED","COMPLETED","CANCELLED"] },
          "creatorId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["title","startDate","endDate","location","category","maxParticipants","creatorId"]
      },
      "EventCreateRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "description": { "type": "string" },
          "startDate": { "type": "string" },
          "endDate": { "type": "string" },
          "location": { "type": "string" },
          "category": { "type": "string" },
          "maxParticipants": { "type": "string" },
          "thumbnailUrl": { "type": "string" }
        },
        "required": ["title","startDate","endDate","location","category","maxParticipants"]
      },
      "ThumbnailUploadRequest": {
        "type": "object",
        "properties": {
          "file": { "type": "string", "format": "binary" }
        },
        "required": ["file"]
      }
    }
  },
  "security": [],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register",
        "description": "Register a new user account. Role can be specified as VOLUNTEER or MANAGER, defaults to VOLUNTEER if not provided.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AuthRegisterRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "User registered successfully" },
                    "user": { "$ref": "#/components/schemas/User" },
                    "accessToken": { "type": "string" },
                    "refreshToken": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "description": "Authenticate user and return access/refresh tokens along with user data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AuthLoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Login successful" },
                    "user": { "$ref": "#/components/schemas/User" },
                    "accessToken": { "type": "string" },
                    "refreshToken": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "Get current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          }
        }
      },
      "put": {
        "summary": "Update current user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateMeRequest" } }
          }
        },
        "responses": { "200": { "description": "Updated" } }
      }
    },
    "/users/me/avatar": {
      "post": {
        "summary": "Upload avatar (Supabase)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UploadAvatarRequest" } }
          }
        },
        "responses": { "200": { "description": "Uploaded" } }
      }
    },
    "/users/{id}/lock": {
      "post": {
        "summary": "Lock/unlock account",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LockRequest" } } }
        },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not Found" } }
      }
    },
    "/events": {
      "get": {
        "summary": "List all events (with optional filters)",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "location", "in": "query", "schema": { "type": "string" } },
          { "name": "q", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "List of events", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } } } } } }
      },
      "post": {
        "summary": "Create a new event (manager/admin only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/EventCreateRequest" } }
          }
        },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Event" } } } } }
      }
    },
    "/events/{id}": {
      "get": {
        "summary": "Get event details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Event details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Event" } } } } }
      },
      "put": {
        "summary": "Update event (owner/admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EventCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated event",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Event" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete event (owner/admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Deleted successfully" } }
      }
    },
    "/events/{id}/submit": {
      "post": {
        "summary": "Submit event for approval",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Submitted for approval", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Event" } } } } }
      }
    },
    "/managers/{id}/events": {
      "get": {
        "summary": "Get events created by a manager",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Manager's events", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } } } } } }
      }
    },
    "/events/upload-thumbnail": {
      "post": {
        "summary": "Upload event thumbnail (Supabase)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": { "$ref": "#/components/schemas/ThumbnailUploadRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/events": {
      "get": {
        "summary": "List events for admin (with optional status filter)",
        "description": "Admin only endpoint to view all events, with optional status filter (e.g., ?status=PENDING_APPROVAL)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "description": "Filter by event status", "schema": { "type": "string" } },
          { "name": "q", "in": "query", "description": "Search query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "description": "Page number for pagination", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "description": "Number of results per page", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "total": { "type": "integer" },
                        "pages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/events/{id}/approve": {
      "post": {
        "summary": "Approve an event",
        "description": "Admin only endpoint to approve an event pending approval",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": { "type": "string", "description": "Optional reason for approval" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event approved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "event": { "$ref": "#/components/schemas/Event" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/events/{id}/reject": {
      "post": {
        "summary": "Reject an event",
        "description": "Admin only endpoint to reject an event pending approval",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": { "type": "string", "description": "Required reason for rejection" }
                },
                "required": ["reason"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event rejected successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "event": { "$ref": "#/components/schemas/Event" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}